esp/design/requirements.txt
paho-mqtt
msgpack